로딩씬 추가 - google 게임에 연결하고 데이터 받아오는데 필요 스테이지 끝까지 남은 거리 보여주는거 추가 Json으로 저장하는 시스템 추가 Google 클라우드에 데이터 전송하는 기능 추가 뭐 많이 추가

+++ Assets/Scripts/playerController.cs
@@ -53,8 +53,10 @@ public class playerController : MonoBehaviour
                 
                 isDoubleJumped = true;
                 anim.SetBool("IsDoubleJump", isDoubleJumped);
-            }            
+            }
         }
+        
+        
     }
 
     void FixedUpdate()
@@ -84,12 +86,10 @@ public class playerController : MonoBehaviour
                     rigied.constraints = RigidbodyConstraints2D.FreezeRotation;
                 }
             }
-
             //게임 시스템에 플레이어 높이 갱신
             tempY = gameObject.transform.position.y;
             GameSystem.playerHeight = tempY;
-            GameSystem.setMaxHeight((int)tempY);
-            
+            GameSystem.setMaxHeight(tempY);
             if(isOnWall){
                 //벽에 부딪혀 있을경우
                 MonkeyOnWall();
